moved light messages from generic directory
