+++ bluetooth_mesh/models/light/lightness.py
@@ -27,7 +27,7 @@ from functools import partial
 from typing import Any, Dict, Iterable, NamedTuple, Optional, Sequence, Tuple, Type
 
 from bluetooth_mesh.models.base import Model
-from bluetooth_mesh.messages.generic.light.lightness import (
+from bluetooth_mesh.messages.light.lightness import (
     LightLightnessOpcode,
     LightLightnessSetupOpcode,
 )
